fix: fix segmentation colormap dtype and cmake custom targets, add results
- Fix numpy uint8 bitwise_or cast error in semantic_segmentation.py
- Fix add_custom_target commands by splitting into multiple COMMANDs
  to avoid pipe/&& being interpreted by make's shell
- Add execution results section to docs (ja/en)

+++ apps/pycoral-examples/src/semantic_segmentation.py
@@ -26,7 +26,7 @@ def create_pascal_label_colormap():
 
     for shift in reversed(range(8)):
         for channel in range(3):
-            colormap[:, channel] |= ((indices >> channel) & 1) << shift
+            colormap[:, channel] |= (((indices >> channel) & 1) << shift).astype(np.uint8)
         indices >>= 3
 
     return colormap
